requirements.txt
Django>=5.0,<6.0
pandas>=2.2,<3.0
requests>=2.31,<3.0
matplotlib>=3.8,<4.0
yfinance>=0.2,<1.0
ccxt>=4.0,<5.0
streamlit>=1.30,<2.0
plotly>=5.20,<6.0
python-dotenv>=1.0,<2.0
dj-database-url>=2.2,<3.0
gunicorn>=22.0,<23.0
whitenoise>=6.7,<7.0